Revert "meta-xilinx: add BSP_XILINX_DERIVED to BBFILES_DYNAMIC"
Issue: LIN1025-1692

This reverts commit 1f418b975b86bf8c2b0300cc97c088e905205d71.

The variable "BBFILES_DYNAMIC" including some bb and bbappend files
will be parsed and its content will be added to "BBFILES". But when it
is parsed, the variables 'MACHINE' and 'BSP_XILINX_DERIVED' are not
ready and this results that BBFILES loses the bb and bbappend files in
"BBFILES_DYNAMIC". So revert the commit b7f164cd ("meta-xilinx: add
BSP_XILINX_DERIVED to BBFILES_DYNAMIC").

(LOCAL REV: NOT UPSTREAM) - WRLinux specific

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>