meta-xilinx-standalone-sdt: drop sanity-meta-xilinx-standalone-sdt check
Issue: LIN1025-5685

The bbclass sanity-meta-xilinx-standalone-sdt is used to check the
condition that meta-xilinx-standalone-sdt layer is included when
the variable XILINX_WITH_ESW is not set to "sdt". If so, a warning
is printed out:

    WARNING: You have included the meta-xilinx-standalone-sdt layer,
    but it has not been enabled using XILINX_WITH_ESW in your configuration.
    Some bbappend files and preferred version setting may not take effect.
    See the meta-xilinx-standalone-sdt README for details.

For WRLinux, the meta-xilinx-standalone-sdt is included by default no
matter XILINX_WITH_ESW is "sdt" or "xsct". So we will drop this check
to silence the warning.

(LOCAL REV: NOT UPSTREAM) - WRLinux specific

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>