meta-xilinx-multimedia: mask gstreamer related bbappend
Issue: LINUXEXEC-37078

The bbappend files for gstreamer are for gstreamer version 1.22.12 in
meta-xilinx-multimedia, but gstreamer package in oe-core has been updated
to 1.24.13. Because we do not support any VPU inside Xilinx SoCs for now,
we mask gstreamer related bbappend.

This patch is to fix the following error:

    ERROR: No recipes in default available for:
	meta-xilinx/meta-xilinx-multimedia/recipes-multimedia/gstreamer/gst-devtools_1.22.%.bbappend

(LOCAL REV: NOT UPSTREAM) - WRLinux Specific

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>