xilinx-standalone: disable stack-protector security cflags for fsbl
Issue: LINUXEXEC-37078

When meta/conf/distro/include/security_flags.inc is included, the option
"-fstack-protector-strong" will be added to ${TARGET_CC_ARCH} to enable
stack overflow security checks.

When building fsbl, the compiler will insert a piece of code which calls
the function "__stack_chk_fail" from the C library "newlib" to the
static objects, for example libxilpm.a, and in the "__stack_chk_fail"
implementation of "newlib", it calls "write" that is a "system call"
not implemented for a bare-metal target "fsbl". When linking these objects
to fsbl.elf, the linker can not find the definition for "write" function.
This triggers the following error:

    tmp-zynqmp-zcu102-sdt-full-cortexa53-fsbl/work/amd_zynqmp-xilinx-elf/fsbl-firmware/2025.2+git/recipe-sysroot-native/usr/bin/aarch64-xilinx-elf/../../libexec/aarch64-xilinx-elf/gcc/aarch64-xilinx-elf/15.2.0/ld: tmp-zynqmp-zcu102-sdt-full-cortexa53-fsbl/work/amd_zynqmp-xilinx-elf/fsbl-firmware/2025.2+git/recipe-sysroot/usr/lib/libc.a(libc_a-stack_protector.o): in function `__stack_chk_fail':
    | /usr/src/debug/newlib/4.4.0+git/newlib/libc/ssp/stack_protector.c:39:(.text+0x58): undefined reference to `write'

The "ESW_CFLAGS" contains the CFLAGS options which will be passed to the
components of fsbl. Add "-fno-stack-protector" to it to disable
stack-protector for fsbl.

(LOCAL REV: NOT UPSTREAM) - WRLinux specific

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>