meta-xilinx: fsbl-firmware: disable -fPIE for fsbl-firmware
Issue: LINUXEXEC-37078

When meta/conf/distro/include/security_flags.inc is included, the option
"-fPIE" will be added to ${TARGET_CC_ARCH}. When the linker tries to
link the components compiled with "-fPIE" enabled, it will create "PHDR"
and ".interp" sections automatically. But it will report the following
error when it finds no way to map the section ".interp" to segment
according to the linker script.

    tmp-zynqmp-zcu102-sdt-full-cortexa53-fsbl/work/amd_zynqmp-xilinx-elf/fsbl-firmware/2025.2+git/recipe-sysroot-native/usr/bin/aarch64-xilinx-elf/../../libexec/aarch64-xilinx-elf/gcc/aarch64-xilinx-elf/14.3.0/ld: zynqmp_fsbl.elf: error: PHDR segment not covered by LOAD segment

The fsbl is a bootloader and no interpreter for it. So we need to
disable -fPIE for fsbl-firmware by using ${SECURITY_NOPIE_CFLAGS} as
SECURITY_CFLAGS.

(LOCAL REV: NOT UPSTREAM) - WRLinux specific

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>